// model/billOfMaterials.ts
/**
 * PingOne Platform API - Management
 * A bare-bones collection for the PingOne API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { BillOfMaterialsProducts } from './billOfMaterialsProducts';

export class BillOfMaterials {
    /**
    * The time the resource was created.
    */
    'createdAt'?: string;
    /**
    * The time the resource was last updated.
    */
    'updatedAt'?: string;
    /**
    * An array that specifies the products associated with this bill of materials
    */
    'products'?: Array<BillOfMaterialsProducts>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "string"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "string"
        },
        {
            "name": "products",
            "baseName": "products",
            "type": "Array<BillOfMaterialsProducts>"
        }    ];

    static getAttributeTypeMap() {
        return BillOfMaterials.attributeTypeMap;
    }
}