// model/inlineObject2.ts
/**
 * PingOne Platform API - Management
 * A bare-bones collection for the PingOne API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class InlineObject2 {
    /**
    * A string that specifies the type of environment to use. Options are PRODUCTION and SANDBOX.
    */
    'type'?: InlineObject2.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "InlineObject2.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return InlineObject2.attributeTypeMap;
    }
}

export namespace InlineObject2 {
    export enum TypeEnum {
        Production = <any> 'PRODUCTION',
        Sandbox = <any> 'SANDBOX'
    }
}